requirements.txt
opencv-python
numpy
tensorflow
torch
torchvision
matplotlib
mediapipe
ultralytics
twilio
python-dotenv
fpdf
smtplib3
